+++ labelbox/data/annotation_types/__init__.py
@@ -18,7 +18,7 @@ from .classification import Dropdown
 from .classification import Radio
 from .classification import Text
 
-from .data import RasterData
+from .data import ImageData
 from .data import TextData
 from .data import VideoData