+++ labelbox/data/annotation_types/data/__init__.py
@@ -1,3 +1,3 @@
-from .raster import RasterData
+from .raster import ImageData
 from .text import TextData
 from .video import VideoData