+++ labelbox/data/serialization/ndjson/metric.py
@@ -1,4 +1,4 @@
-from labelbox.data.annotation_types.data import TextData, RasterData
+from labelbox.data.annotation_types.data import TextData, ImageData
 from labelbox.data.annotation_types.metrics import ScalarMetric
 from labelbox.data.serialization.ndjson.base import NDJsonBase
 from typing import Union
@@ -13,7 +13,7 @@ class NDDataRowScalarMetric(NDJsonBase):
     @classmethod
     def from_common(
             cls, metric: ScalarMetric,
-            data: Union[TextData, RasterData]) -> "NDDataRowScalarMetric":
+            data: Union[TextData, ImageData]) -> "NDDataRowScalarMetric":
         return NDDataRowScalarMetric(uuid=metric.extra.get('uuid'),
                                      metric_value=metric.value,
                                      data_row={'id': data.uid})
@@ -27,7 +27,7 @@ class NDMetricAnnotation:
 
     @classmethod
     def from_common(cls, annotation: ScalarMetric,
-                    data: Union[TextData, RasterData]) -> NDDataRowScalarMetric:
+                    data: Union[TextData, ImageData]) -> NDDataRowScalarMetric:
         return NDDataRowScalarMetric.from_common(annotation, data)